Add Mission.PlayingSurface
