+++ lib/chukinas_web/live/dreadnought_live.html.leex
@@ -11,18 +11,23 @@
   }
 </style>
 
+<%# TODO delete this and route %>
 <%# World: Welcome Demo Screen %>
 <%= if @live_action == :welcome, do: 
   template "welcome_animation.html", socket: @socket %>
 
 <%# World: Actual Play %>
 <%= if @live_action != :welcome do %>
-  <%= live_component @socket, Dreadnought.StaticWorldComponent, id: :grid_lab, mission: @mission do %>
-    <%= live_component @socket,
+<%= live_component @socket, 
+    Dreadnought.StaticWorldComponent, 
+    id: :grid_lab, 
+    mission_static: @mission_static do %>
+  <%# TODO remove the ChukinasWeb %>
+  <%= live_component @socket,
       ChukinasWeb.Dreadnought.DynamicWorldComponent,
       id: :dynamic_world,
-      mission: @mission %>
-  <% end %>
+      mission_player: @mission_player %>
+<% end %>
 <% end %>
 
 <%= live_component @socket, Dreadnought.MenuBarComponent, id: :menu_bar %>
